fix: Resolve Flutter compilation errors
- Fix indentation in my_orders_screen.dart
- Convert rating from double to int for RatingWidget
- Change DateRange to DateTimeRange in photographer_payments_screen
- Remove unnecessary toDouble() call in my_reviews_screen

+++ front/lib/screens/photographer/my_reviews_screen.dart
@@ -143,8 +143,7 @@ class _PhotographerReviewsScreenState extends State<PhotographerReviewsScreen> {
                                 mainAxisAlignment:
                                     MainAxisAlignment.spaceBetween,
                                 children: [
-                                  RatingWidget(
-                                      rating: review.rating.toDouble()),
+                                  RatingWidget(rating: review.rating),
                                   Text(
                                     'Заказ ID: ${review.orderId.substring(review.orderId.length - 6)}',
                                     style: TextStyle(
